Extract lexical name path construction
