+++ crates/ty_python_semantic/src/lib.rs
@@ -56,6 +56,7 @@ pub use types::{
 mod db;
 mod dunder_all;
 mod fixes;
+mod lexical_name_path;
 pub mod lint;
 pub(crate) mod place;
 pub(crate) mod place_load;
